refactor repositories, convert loader to repository, allow group of dataRepositories, fix contexts

+++ internal/repository/influx.go
@@ -7,7 +7,6 @@ import (
 	"github.com/influxdata/influxdb-client-go/v2"
 	"github.com/influxdata/influxdb-client-go/v2/api"
 	"github.com/oitimon/day-ahead-prices-notificator/internal/config"
-	"github.com/oitimon/day-ahead-prices-notificator/internal/loader"
 	"github.com/shopspring/decimal"
 	"log"
 	"time"
@@ -16,23 +15,28 @@ import (
 const measurementPrices = "prices"
 const fieldNamePrice = "price"
 const tagCurrency = "currency"
+const currencyValue = "EUR"
 
 type Influx struct {
-	ctx       context.Context
-	cfg       *config.Influx
-	client    influxdb2.Client
-	writeAPI  api.WriteAPI
-	queryAPI  api.QueryAPI
-	deleteAPI api.DeleteAPI
-	ldr       loader.Loader
+	cfg               *config.Influx
+	client            influxdb2.Client
+	writeAPI          api.WriteAPI
+	queryAPI          api.QueryAPI
+	deleteAPI         api.DeleteAPI
+	prev              Data
+	deleteBeforeWrite bool
 }
 
-func NewInflux(ctx context.Context, cfg *config.Influx, ldr loader.Loader) *Influx {
-	inf := &Influx{
-		ctx:    ctx,
+func NewInflux(ctx context.Context, cfg *config.Influx, prev Data) (inf *Influx, err error) {
+	inf = &Influx{
 		cfg:    cfg,
 		client: influxdb2.NewClient(cfg.Url, cfg.Token),
-		ldr:    ldr,
+		prev:   prev,
+		// You can set deleteBeforeWrite to true if you want to delete the data before writing it.
+		// This is useful if you want to overwrite the data in the database.
+		// But take into account it will delete all fields with the same measurement and tags.
+		// It's not recommended to use it (only for debugging purposes).
+		deleteBeforeWrite: false,
 	}
 	inf.writeAPI = inf.client.WriteAPI(cfg.Orgname, cfg.Bucket)
 	inf.queryAPI = inf.client.QueryAPI(cfg.Orgname)
@@ -41,30 +45,26 @@ func NewInflux(ctx context.Context, cfg *config.Influx, ldr loader.Loader) *Infl
 	errorsCh := inf.writeAPI.Errors()
 	go func() {
 		for err := range errorsCh {
-			log.Printf("Write error: %v", err)
+			log.Printf("Influxdb Writer error: %v", err)
 		}
 	}()
 
 	// Close the client when the context is done.
 	go func() {
+		// Main server context, to close the client when the server is closed.
 		<-ctx.Done()
 		log.Println("Closing InfluxDB client...")
 		inf.client.Close()
 	}()
 
-	return inf
+	return
 }
 
-func (inf *Influx) Data() Data {
-	return inf
+func (*Influx) IsFinal() bool {
+	return false
 }
 
-func (inf *Influx) Bytes() Bytes {
-	// @TODO: Implement this method
-	return nil
-}
-
-func (inf *Influx) Get(startDate time.Time) (prices []decimal.Decimal, err error) {
+func (inf *Influx) Get(ctx context.Context, startDate time.Time) (prices []decimal.Decimal, err error) {
 	// Read from DB first.
 	endDate := time.Date(startDate.Year(), startDate.Month(), startDate.Day(), 23, 59, 59, 0, startDate.Location())
 	query := fmt.Sprintf(`
@@ -80,7 +80,7 @@ func (inf *Influx) Get(startDate time.Time) (prices []decimal.Decimal, err error
 	)
 	log.Printf("Query to Influx: %s\n", query)
 	// We create a sprintf-query as QueryWithParams is not working properly in the Go client.
-	result, err := inf.queryAPI.Query(inf.ctx, query)
+	result, err := inf.queryAPI.Query(ctx, query)
 	if err != nil {
 		err = errors.New("error querying data from Influx: " + err.Error())
 		return
@@ -103,28 +103,31 @@ func (inf *Influx) Get(startDate time.Time) (prices []decimal.Decimal, err error
 	}
 
 	if count <= 0 {
-		prices, err = inf.loadData(startDate)
+		prices, err = inf.loadData(ctx, startDate)
 	}
 	return
 }
 
-func (inf *Influx) loadData(startDate time.Time) (prices []decimal.Decimal, err error) {
-	// Clear existing data.
-	endDate := time.Date(startDate.Year(), startDate.Month(), startDate.Day(), 23, 59, 59, 0, startDate.Location())
-	if err = inf.deleteAPI.DeleteWithName(
-		inf.ctx,
-		inf.cfg.Orgname,
-		inf.cfg.Bucket,
-		startDate,
-		endDate,
-		fmt.Sprintf(`_measurement="%s"`, measurementPrices),
-	); err != nil {
-		err = errors.New("error deleting old data from Influx: " + err.Error())
-		return
+func (inf *Influx) loadData(ctx context.Context, startDate time.Time) (prices []decimal.Decimal, err error) {
+	if inf.deleteBeforeWrite {
+		log.Printf("WARNING: Deleting data from InfluxDB for %s\n", startDate.Format("2006-01-02"))
+		// Clear existing data (all fields).
+		endDate := time.Date(startDate.Year(), startDate.Month(), startDate.Day(), 23, 59, 59, 0, startDate.Location())
+		if err = inf.deleteAPI.DeleteWithName(
+			ctx,
+			inf.cfg.Orgname,
+			inf.cfg.Bucket,
+			startDate,
+			endDate,
+			fmt.Sprintf(`_measurement="%s"`, measurementPrices),
+		); err != nil {
+			err = errors.New("error deleting old data from Influx: " + err.Error())
+			return
+		}
 	}
 
 	// Fetch data from the loader.
-	prices, err = inf.ldr.Fetch(startDate)
+	prices, err = inf.prev.Get(ctx, startDate)
 	if err != nil {
 		err = errors.New("error fetching prices: " + err.Error())
 		return
@@ -134,7 +137,7 @@ func (inf *Influx) loadData(startDate time.Time) (prices []decimal.Decimal, err 
 	for i := 0; i < len(prices); i++ {
 		p := influxdb2.NewPoint(
 			measurementPrices,
-			map[string]string{"currency": tagCurrency},                         // tags
+			map[string]string{tagCurrency: currencyValue},                      // tags
 			map[string]interface{}{fieldNamePrice: prices[i].InexactFloat64()}, // fields
 			startDate.Add(time.Hour*time.Duration(i)),                          // timestamp
 		)
